Enable sleep timer on startup
To preserve the battery health, start the sleep timer when the device boots.

If the device doesn't receive a keypress after 30mins it will power off.

If you plan to use rockbox unattended (e.g. a party) or to listen to an album sequentially remember to disable the sleep timer in settings for this specific occasion.

9 times out of 10 the behaviour of this commit is what I want - prioritising battery health - as I frequently skip tracks and this will restart the sleep timer.